// packages/wscapi/types/Wscapi.ts
import type { Pointer } from 'bun:ffi';

export type { DWORD, HANDLE, HRESULT, LPVOID, NULL, PHANDLE, PVOID } from '@bun-win32/core';

export enum WSC_SECURITY_PROVIDER {
  WSC_SECURITY_PROVIDER_ALL = 0x7f,
  WSC_SECURITY_PROVIDER_ANTISPYWARE = 0x08,
  WSC_SECURITY_PROVIDER_ANTIVIRUS = 0x04,
  WSC_SECURITY_PROVIDER_AUTOUPDATE_SETTINGS = 0x02,
  WSC_SECURITY_PROVIDER_FIREWALL = 0x01,
  WSC_SECURITY_PROVIDER_INTERNET_SETTINGS = 0x10,
  WSC_SECURITY_PROVIDER_NONE = 0x00,
  WSC_SECURITY_PROVIDER_SERVICE = 0x40,
  WSC_SECURITY_PROVIDER_USER_ACCOUNT_CONTROL = 0x20,
}

export enum WSC_SECURITY_PROVIDER_HEALTH {
  WSC_SECURITY_PROVIDER_HEALTH_GOOD = 0,
  WSC_SECURITY_PROVIDER_HEALTH_NOTMONITORED = 1,
  WSC_SECURITY_PROVIDER_HEALTH_POOR = 2,
  WSC_SECURITY_PROVIDER_HEALTH_SNOOZE = 3,
}

export type LPTHREAD_START_ROUTINE = Pointer;
export type PLPWSTR = Pointer;
export type PPVOID = Pointer;
export type PWSC_SECURITY_PROVIDER_HEALTH = Pointer;
export type REFCLSID = Pointer;
export type REFIID = Pointer;